Improve code quality throughout codebase
There are several methods that contain deep nesting with arrowhead
style code. This makes it hard to keep track of the logic. The happy
paths are also unclear.

This reduces readability and code understanding throughout the codebase.

Thus, we add guard clauses and apply SLAP (single level of abstraction
per function). We also deleted dead code and minimised code duplication.